+++ EDTracking/EDRoute.cs
@@ -12,9 +12,9 @@ namespace EDTracking
     {
         public string Name { get; set; } = null;
         public List<EDWaypoint> Waypoints { get; set; } = null;
-        private double _totalWaypointDistance = 0;
-        private List<double> _waypointDistances = new List<double>();
-        private List<double> _distanceLeftAtWaypoint = new List<double>();
+        private decimal _totalWaypointDistance = 0;
+        private List<decimal> _waypointDistances = new List<decimal>();
+        private List<decimal> _distanceLeftAtWaypoint = new List<decimal>();
         private int _lastWaypointCount = 0;
         private string _saveFilename = "";
 
@@ -41,8 +41,8 @@ namespace EDTracking
             if (Waypoints.Count == _lastWaypointCount)
                 return;
 
-            _distanceLeftAtWaypoint = new List<double>();
-            _waypointDistances = new List<double>();
+            _distanceLeftAtWaypoint = new List<decimal>();
+            _waypointDistances = new List<decimal>();
             _totalWaypointDistance = 0;
             if (Waypoints.Count < 2)
                 return;
@@ -58,7 +58,7 @@ namespace EDTracking
                 _distanceLeftAtWaypoint.Add(_distanceLeftAtWaypoint[i] - _waypointDistances[i]);
         }
 
-        public double TotalDistanceLeftAtWaypoint(int WaypointIndex)
+        public decimal TotalDistanceLeftAtWaypoint(int WaypointIndex)
         {
             CalculateDistances();
             if (WaypointIndex < _distanceLeftAtWaypoint.Count)
